improve paragraphs
