+++ src/components/Paragraph/Paragraph.css
@@ -1,0 +1,3 @@
+.row {
+  width: 610px;
+}
