+++ src/components/Paragraph/ParagraphText.css
@@ -1,0 +1,24 @@
+.column.text {
+  width: var(--flat-image-width);
+  float: right;
+  margin: 0;
+  padding: 0;
+  font-size: 14px;
+  font-family: var(--font-poppins-light);
+  white-space: pre-wrap;
+  text-align: justify;
+  letter-spacing: 1;
+}
+
+.column.text p {
+  margin: 0;
+}
+
+.text p span {
+  font-family: var(--font-poppins-bold);
+}
+
+.text p span a {
+  text-decoration: none;
+  color: unset;
+}
